requirements.txt
pandas
streamlit
feedparser
python-dotenv
openai
gptcache
beautifulsoup4
lxml
SQLAlchemy
psycopg2-binary
plotly
yfinance>=0.2.3
PyYAML
holidays
pytz
holidays
scikit-learn
joblib
spacy
python-fasthtml
uvicorn
aiohttp
werkzeug
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl